Set up hands-free GTID replication on replica startup
By using the GTID replication available in MySQL 5.7, we can avoid
having the replica `onStart` manually probing inside the primary in
order to find out the binlog position. Instead the DB will use GTIDs
to auto-configure this position.

This will work for cases where we're first starting up a cluster.
For starting up instances on an existing cluster with signficant
data, we'll want to add the ability to migrate a dump of the existing
data first.