Write backups to Manta and use them to bootstrap replication
Once the primary is set up, have it immediately push a backup to Manta
and write a key to Consul telling what the last backup was. We also write
a TTL to Consul that will timeout once it's time to write a new backup.

When replicas come up, they check in with Consul to get the most recent
backup, pull that file from Manta, and then bootstrap replication from
that point.

change details:
- Switch to Percona base
- Use Percona xtrabackup for taking snapshots of primary to bootstrap GTID
  replication for subsequent instances.
- Use `repl` (dedicated user) for set up and management of all replication
  and backup operations.

major TODOs:
- testing that we can fail-out primary and reboot/start a replica to take
  over.
- testing that replication bootstraps correctly with no window of lost
  transactions (has been tested to work in trivial cases).