Have primary snapshot whenever binlog rotates
Write the binlog filename to Consul whenever there's a snapshot. On health
check, if the binlog file name changes from what's in Consul then it's been
rotated and we do a new snapshot.

This commit also combines the behaviors of the primary and standby into a
single behavior, but provides an optional override via USE_STANDBY env var.