Use Consul sessions to lock leader election for `onChange` handler
(First pass-thru of this.)

The key we use to mark the primary in Consul is now locked by a session
with TTL. The primary updates this TTL with each pass thru the `health`
handler. If the primary becomes unhealthy, the replica(s) will try to
get this lock in their `onChange` handler. The winner will become the
new replica and the old replicas will update their replication config
to point to it.