Update MySQLNode state prior to passing control to on_change
When the MySQLNode instance is instantiated it doesn't know if the node
is primary or not. In the on_change handler we check to see if the node
is primary but did not first update the node from Consul so the check
always failed and the primary would execute the on_change behaviors we
expect from the replicas. This would be harmless but races with health
handler behaviors during the initial snapshot.