Make sure we create the backup lock file in create_snapshot
If we don't make sure we create the file here we'll throw an IOError and catch
it in the except block. This is ok in `is_backup_running` because if the file
doesn't even exist we're obviously not running the backup either. But we need
to make sure we create the file in `create_snapshot` or we end up bailing out
too early.