Add a file lock around creation of a snapshot (#20)
* Add a file lock around creation of a snapshot

Because we set the "binlog isn't stale" flag at the end of a backup, a newly
started cluster will spawn many `create_snapshot` processes even though we're
locking the BACKUP_TTL in Consul. This adds a local file lock that prevents
the node from running multiples.

Also removed an unnecessary reliance on the repl user having access to the
end-user's data DB just to run health checks.

* Make sure we create the backup lock file in create_snapshot

If we don't make sure we create the file here we'll throw an IOError and catch
it in the except block. This is ok in `is_backup_running` because if the file
doesn't even exist we're obviously not running the backup either. But we need
to make sure we create the file in `create_snapshot` or we end up bailing out
too early.