Move DB initialization into the health check
MySQL must be running in order to execute most of our setup behavior
so we're just going to make sure the directory structures are in
place and then let the first health check handler take it from there.

The happy path on the health check is to check a lock file against the
node state and immediately return if we discover the lock exists.
Otherwise, we bootstrap the instance.