Move get_snapshot into pre_start
Restoring from snapshot has to happen before we start mysqld, so if we have
a clean start then we need to check in with Consul during startup to see if
there's a snapshot available.

Added debug trace logging and environment variable cleanup functions to
assist with development and otherwise make life sane.