Break on_change on ER_SLAVE_MUST_STOP
There is a race between the onChange and health handler around stopping
replication while we figure out which instance is primary. These leaves
the onChange handler in an infinite loop because we previously assumed
that the only errors we'd get were from the session lock.

This commit breaks the loop when error code 1198 ER_SLAVE_MUST_STOP is
received from mysqld, which says that replication has already been started.