Record timestamp for last snapshot rather than overloading service
When we registered a check for the snapshot, it was not being bound to a
particular service and so when it failed the check the entire node was being
marked as unhealthy. Instead of further overloading the Consul services concept,
we'll just write a key with the last snapshot's timestamp to Consul and then
compare the timestamps against the current time and the BACKUP_TTL value.